+++ src/Command/Strategy/AbstractFactory.php
@@ -28,6 +28,8 @@ namespace PhpDA\Command\Strategy;
 use Fhaculty\Graph\Graph;
 use PhpDA\Layout\Builder;
 use PhpDA\Layout\GraphViz;
+use PhpDA\Layout\Helper\CycleDetector;
+use PhpDA\Layout\Helper\DependencyMapGenerator;
 use PhpDA\Layout\Helper\GroupGenerator;
 use PhpDA\Parser\AnalyzerFactory;
 use PhpDA\Plugin\FactoryInterface;
@@ -60,7 +62,9 @@ abstract class AbstractFactory implements FactoryInterface
      */
     protected function createGraphBuilder()
     {
-        return new Builder(new GraphViz(new Graph), new GroupGenerator);
+        $cycleDetector = new CycleDetector(new DependencyMapGenerator);
+
+        return new Builder(new GraphViz(new Graph), new GroupGenerator, $cycleDetector);
     }
 
     /**