+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## dev-master
 
+- Change colors used in layout to 255 palette
+- Add Cycle Hilighting feature
 - Remove edge concentration in GraphViz to improve usability with ReferenceValidator Feature
 - Add ReferenceValidator Feature